♻️(typed-value): Split long method to small methods.
